✨ dependency|bump => extend config. Add new section "description" in example
It makes examples more verbose. In some case I(as a config writer) want to emphasize reader's attention on something in PR

for example, I have 3 examples in config. 2 of them are easy and same, and 3d has an additional action to do, and I want to highlight this nuance

More verbose and understandable explanation - more easy to deal with it for final user